renamed comment component
